requirements.txt
numpy
pyyaml
h5py
torch
torchvision
matplotlib
tqdm
